refactor random message handling
