+++ events/messageCreate.ts
@@ -1,26 +1,5 @@
 import { Client, Message } from "discord.js";
-
-import fs from "fs";
-import readline from "readline";
-
-async function processFile() {
-	const fileStream = fs.createReadStream("./data/axer.txt");
-	const returnArray = [];
-	const rl = readline.createInterface({
-		input: fileStream,
-		crlfDelay: Infinity,
-	});
-	// Note: crlfDelay is to recognize all instances of CR LF
-	// ('\r\n') in the .txt as a single line break.
-
-	for await (const line of rl) {
-		// Each line in the .txt will be successively available here as `line`.
-		//TODO: fetch lines as template literals instead of strings with quotes
-		returnArray.push(line);
-	}
-
-	return returnArray;
-}
+import processFile from "../utils/messages/randomMessage";
 
 export default {
 	name: "messageCreate",
@@ -36,7 +15,7 @@ export default {
 				message.content.toUpperCase().includes("AXER") ||
 				message.mentions.has(bot_user)
 			) {
-				processFile().then(function (data) {
+				processFile("./data/axer.txt").then(function (data) {
 					let quotes = data;
 					let randomQuote =
 						quotes[Math.floor(Math.random() * quotes.length)];
